backend/requirements.txt
fastapi
uvicorn[standard]

pydantic
pydantic-settings
python-dotenv

langchain
langchain-community
langchain-core
langchain-text-splitters

langgraph
langgraph-prebuilt
langgraph-sdk

langfuse
litellm
openai

chromadb
pypdf

numpy
requests
httpx

python-multipart
tiktoken
tenacity
